PluralSight game project
This is the final project of web development with the Django Tutorial
presented at pluralsight. Please, note that the course was taught using
django 1.6 and I worked by myself trying to figure out how could I write
this project with the newest version(1.10 by now).

The webpage result is different than the one taught in the course. But
it is a good comparisson. Guys, fell free to give me hints and ideas.